requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
pkginfo==1.4.2
requests==2.19.1
requests-toolbelt==0.8.0
tqdm==4.23.4
twine==1.11.0
urllib3==1.23
websocket-client==0.40.0
